Split various glyph structs in their own files
Also rename ContourPoint to Point.
